move existing tests to __tests__ directory
